Cleaned up directories, moved code
